requirements.txt
aiohttp>=3.10
shioaji>=1.2
sqlalchemy[asyncio]>=2.0
asyncpg>=0.29
websockets>=12.0
python-dotenv>=1.0
pyinstaller>=6.0
